Consolidate OpenAI logic into it's new ModelProviders file. Move general concepts to LLMProviders.ts
