+++ src/components/Chat/NewChat.tsx
@@ -70,7 +70,7 @@
 // import { fetchContexts } from '~/pages/api/getContexts'
 // import { useUser } from '@clerk/nextjs'
 // import { extractEmailsFromClerk } from '../UIUC-Components/clerkHelpers'
-// import { type OpenAIModelID, OpenAIModels } from '~/types/openai'
+// import { type OpenAIModelID, OpenAIModels } from '~/utils/modelProviders/openai'
 // import Navbar from '../UIUC-Components/Navbar'
 // import TopBarInChat from '../Chatbar/TopBarInChat'
 // // import { MainPageBackground } from '../UIUC-Components/MainPageBackground'
@@ -173,7 +173,7 @@ import { useRouter } from 'next/router'
 import { fetchContexts } from '~/pages/api/getContexts'
 import { useUser } from '@clerk/nextjs'
 import { extractEmailsFromClerk } from '../UIUC-Components/clerkHelpers'
-import { type OpenAIModelID, OpenAIModels } from '~/types/openai'
+import { type OpenAIModelID, OpenAIModels } from '~/utils/modelProviders/openai'
 import TopBarInChat from '../Chatbar/TopBarInChat'
 // import { MainPageBackground } from '../UIUC-Components/MainPageBackground'
 import { notifications } from '@mantine/notifications'
